requirements.txt
Django==4.1.2
django_taggit==3.1.0
